Script works, need to check why kubectl version check does not work

+++ p3/install_k3d.sh
@@ -5,37 +5,33 @@ K3D_VERSION=$(k3d version 2>/dev/null)
 KUBECTL_VERSION=$(kubectl --version 2>/dev/null)
 
 set -euo pipefail 
-if [[ -n $K3D_VERSION ]]; then
-	echo "INSTALL_K3D: K3D already installed."
-	exit 0
-fi
 
 if [[ -z $DOCKER_VERSION ]]; then
 	echo "INSTALL_K3D: Installing Docker..."
 	echo "INSTALL_K3D: Removing conflicting packages..."
-	for pkg in  docker.io docker-doc docker-compose docker-compose-v2 podman-docker containerd runc; do sudo apt-get remove -y "$pkg"; done
+	for pkg in docker.io docker-doc docker-compose podman-docker containerd runc; do sudo apt-get remove $pkg; done
 	echo "INSTALL_K3D: Setting up Docker's apt repos..."
-	sudo apt-get update
+	sudo apt-get update -y
 	sudo apt-get install ca-certificates curl -y
 	sudo install -m 0755 -d /etc/apt/keyrings
-	sudo curl -fsSL https://download.docker.com/linux/ubuntu/gpg -o /etc/apt/keyrings/docker.asc
+	sudo curl -fsSL https://download.docker.com/linux/debian/gpg -o /etc/apt/keyrings/docker.asc
 	sudo chmod a+r /etc/apt/keyrings/docker.asc
 	echo \
-	   "deb [arch=$(dpkg --print-architecture) signed-by=/etc/apt/keyrings/docker.asc] https://download.docker.com/linux/ubuntu \
-	     $(. /etc/os-release && echo "${UBUNTU_CODENAME:-$VERSION_CODENAME}") stable" | \
+		"deb [arch=$(dpkg --print-architecture) signed-by=/etc/apt/keyrings/docker.asc] https://download.docker.com/linux/debian \
+		$(. /etc/os-release && echo "$VERSION_CODENAME") stable" | \
 	sudo tee /etc/apt/sources.list.d/docker.list > /dev/null
-	sudo apt-get update
+	sudo apt-get update -y
 	echo "INSTALL_K3D: Installing Docker package..."
-	sudo apt-get install docker-ce docker-ce-cli containerd.io docker-buildx-plugin docker-compose-plugin
+	sudo apt-get install docker-ce docker-ce-cli containerd.io docker-buildx-plugin docker-compose-plugin -y
 	echo "INSTALL_K3D: Testing Docker install..."
 	sudo docker run hello-world
 fi
 
 if [[ -z $KUBECTL_VERSION ]]; then
 	echo "INSTALL_K3D: Downloading kubectl..."
-	curl -LO  "https://dl.k8s.io/release/$(curl -L -s https://dl.k8s.io/release/stable.txt)/bin/linux/arm64/kubectl"
+	curl -LO  "https://dl.k8s.io/release/$(curl -L -s https://dl.k8s.io/release/stable.txt)/bin/linux/amd64/kubectl"
 	echo "INSTALL_K3D: Downloading kubectl checksum..."
-	curl -LO "https://dl.k8s.io/release/$(curl -L -s https://dl.k8s.io/release/stable.txt)/bin/linux/arm64/kubectl.sha256"
+	curl -LO "https://dl.k8s.io/release/$(curl -L -s https://dl.k8s.io/release/stable.txt)/bin/linux/amd64/kubectl.sha256"
 	echo "INSTALL_K3D: Validating kubectl install file"
 	echo "$(cat kubectl.sha256)  kubectl" | sha256sum --check
 	echo "INSTALL_K3D: Installing kubectl..."
@@ -44,9 +40,15 @@ if [[ -z $KUBECTL_VERSION ]]; then
 	kubectl version --client
 fi
 
-echo "INSTALL_K3D: Installing k3d..."
-curl -s https://raw.githubusercontent.com/k3d-io/k3d/main/install.sh | bash
+if [[ -n $K3D_VERSION ]]; then
+	echo "INSTALL_K3D: K3D already installed."
+	exit 0
+fi
+
 echo "INSTALL_K3D: Cleaning up"
 rm -f kubectl
 rm -f kubectl.sha256
+
+echo "INSTALL_K3D: Installing k3d..."
+curl -s https://raw.githubusercontent.com/k3d-io/k3d/main/install.sh | bash
 echo "INSTALL_K3D: Done!"
